Extract rigid body quantization
