+++ rocketsim/src/shared/mod.rs
@@ -1,6 +1,7 @@
 mod aabb;
 pub(crate) mod bvh;
 mod quad_ray;
+pub(crate) mod quantize;
 pub mod rand;
 pub mod rsmath;
